feat: Add JS to disable "Next Post" for 5 seconds
This commit adds some basic JavaScript that will disable the "Next
Post" for 5 seconds after page load.

This complements the CSS change in 837af0 which disables the button
visually for 2 seconds, but does not stop users from pushing it. I have
increased the CSS value from that patch to 5 seconds to match the JS.

Naturally this patch will work just fine if the user disables JS, but
the disabled button will be purely visual.

I have added the JS to the bottom of the index.html template instead of
a separate file as there is no other JS on the page, so it seems a bit
silly to force the browser to make another HTTP request just for a few
lines of JS.