Update smallweb.txt
Added the blog "https://splendide-mendax.com/rss.xml", as I think it matches the criteria of what we're looking for and – while the last post is slightly outside the 7 day window – posts are frequent and substantial enough to justify it. The Bumblebee framework the author hosts alone makes it worth including in the list, and has been of great personal use to me.