Manually resolve new domain
This domain was unresolveable. But since this is an .ua domain and there
is an ongoing war in ukraine, I thought maybe some more research might
be good. archive.org was helpful:

https://web.archive.org/web/20250314025043/http://www.blog.debiania.in.ua/

The last snapshot of this page in the archive shows a redirect, the
owner just seems to have switched domains. I checked the new page and it
does indeed look like the previous one. -> Migrate this